+++ aiohttp/web_server.py
@@ -9,7 +9,7 @@ from .server import ServerHttpProtocol
 from .web_exceptions import HTTPException, HTTPInternalServerError
 from .web_reqrep import BaseRequest
 
-__all__ = ('RequestHandler', 'WebServer')
+__all__ = ('RequestHandler', 'Server')
 
 
 class RequestHandler(ServerHttpProtocol):
@@ -104,9 +104,11 @@ class RequestHandler(ServerHttpProtocol):
         self._request = None
 
 
-class WebServer:
+class Server:
 
     def __init__(self, handler, *, request_factory=None, loop=None, **kwargs):
+        if loop is None:
+            loop = asyncio.get_event_loop()
         self._handler = handler
         self._request_factory = request_factory or self._make_request
         self._loop = loop